server/requirements.txt
bottle==0.12.16
requests==2.21.0
cherrypy==8.9.1
pymongo==3.7.2
gevent==1.4.0
gevent-websocket==0.10.1
pytest==5.0.1
tinydb==3.15.2
slackclient==2.5.0